Use bitfield for win mode instead of manual masking
The bitfield syntax is much better than the mix of bitwise ops and
macros that comes with manual masking. We do not care about bit order
either, so this is perfectly fine portability-wise. As a bonus, this
allows us to store the mouse report mode in a way that reflects that
the four modes are mutually exclusive.

Because enum win_mode and xmode() are gone, we need a new way of tty.c
mutating the win mode in win.c. Currently, we've simply made win global,
but perhaps we should instead expose some setter functions or something.

This change is incomplete. Other code needs to be updated.